Remove all mentions and usage of the `BIN` role (#175)
* Remove all instances of the BIN role

Remove all places where the BIN role is mentioned or used as we have
decided that this role is redundant and its purpose can be taken by
the top-level "targets" role.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Add missing confirmation question in the docs

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Update the ceremony tests do not test the BIN role

The ceremony tests were updated so that the tests behave as the latest
version of the bootstrap ceremony.
The changes are mostly about removing unnecessary inputs.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Fix typo

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Update examples to use hash bin delegation

Change the examples to showcase how hash bin delegation works.

Additionally, improve the question regarding the amount of hash bin
delegations.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Change the order of the hash bin question

Move the hash bin delegation question right after we have shown the
example for an organization that uses hash bin delegation.
This seems more logical to me as we have just explained what hash
bin delegation is.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

* Change example target paths and recalculate them

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>